+++ cursepy/__init__.py
@@ -4,10 +4,10 @@ cursepy init file.
 
 # Import the necessary components:
 
-from cursepy.wrapper import CurseClient
+from cursepy.wrapper import CurseClient, MinecraftWrapper
 
 
 # Define some metadata here:
 
-__version__ = '1.2.0'
+__version__ = '1.3.0'
 __author__ = 'Owen Cochell'